perf(S13): P7 HNSW seqlock + P8 中批 6/14（短语 thread_local / 稀有词驱动 / fuzzy 并行 / on_delete 短路 / C get 零拷贝 / FOR 64-bit 解包）
接 D9（5df2ff1）。本批关 S13-C 群 P7 + P8 大头（6/14 项）——纯性能/并发
路径，位级行为不变；perf 提升靠 bench 量化（P7 新增并发基线，P8 各
项延用既有基准）。

【中·P7】HNSW per-node 自旋锁 → seqlock

`copy_neighbors` 此前对 per-node 自旋锁做 exchange（写操作）——HNSW
流量高度偏向 hub 节点 ⟹ 并发查询时锁缓存行核间乒乓（多核 scaling
不可见/退化）。HNSW 写者单线程（reducer 上下文）⟹ 写-写互斥**不需
要**，只缺发布协议。

修：per-node 锁从 `atomic<uint8_t>` 自旋锁改为 `atomic<uint32_t>`
**seqlock**（u32 防 ABA 回绕）——

* **写者**（reducer 线程）：`seq_write_begin`（fetch_add→奇，release
  fence）→ `atomic_ref<u32>` relaxed 写 adj → `seq_write_end`（fetch_add
  →偶，release 发布）。全程在单线程上下文执行，无写-写竞态。
* **读者**：`acquire` 读 seq（奇则 cpu_pause 退避）→ `atomic_ref<u32>`
  relaxed 读 count → cap 钳制（防 torn count 越界）→ relaxed 读
  n 个 adj → `acquire fence` → 复读 seq 一致才采信，否则丢弃重读。
  **零共享行写**——hub 节点并发查询不再核间乒乓。

数据字全走 `std::atomic_ref`（UB-free 且 TSan 干净——无非原子冲突
访问）。节点发布（`count release`）前的 adj 初始化是单线程预发布阶
段的普通写，与发布后的 `atomic_ref` 访问经 happens-before 分隔，不
构成并发混用。

新增 bench `BM_Hnsw_SearchConcurrent`（100k × ef64，Threads 1/4/8，
UseRealTime）——验证扩展性：1→4 线程延迟持平（147→135µs），8 线程
CPU 时间不涨（无锁争用证据）。HNSW 15/15（含 recall）通过。

【中·P8 6/14】杂项性能批

* **P8.1：短语打分 `other_pos` thread_local**（`inverted.cpp`）—— 候选
  评分闭包内每 doc `std::vector<const vector<u32>*> other_pos(tps.size())`
  此前每次栈/堆分配，在 `tbb::parallel_for` 内 → 分配器争用热点。
  改 `thread_local` 复用 + `.assign(tps.size(), nullptr)`。
* **P8.2：短语驱动词选最稀有**（`inverted.cpp:search_phrase_impl`）——
  候选枚举从恒 `tps[0]` 改为 argmin(`pl.items.size()`) 的最稀有词。
  "the quantum" 不再遍历 "the" 的大表（"the" 百万级、"quantum" 几千）。
  候选集（交集）不变；两列表均按 ord 升序 ⟹ 推入 (score, ord) 序一
  致，top-k 含平分决策**逐字节同果**。idf 仍取 first term 的 live_df
  （BM25 公式不变）。drv==0 时复用 `cand_live` 免二次 live gather。
* **P8.4：`search_fuzzy` 并行化**（`inverted.cpp:search_fuzzy`）——
  镜像 `search_wildcard` 的 `tbb::parallel_reduce` 结构（按 shard 分
  区互不重叠）：每分片线性扫排序 vocab_，命中词 `snapshot_flat` 出
  posting；reduce 阶段 `insert` 合流。MyersMatcher::within 是 const
  纯计算（per-查询词只读 Peq 表 + 局部 DP），跨线程并发安全。`shard
  .inverted.find` 仍 per-shard mutex（与 wildcard 一致）。**评分一
  致性**：tps 集合相同、每 term 恰一份、BM25 贡献按 term 求和（顺序
  无关）——位级不变。
* **P8.5：`on_delete` 空缓存跳过重分词**（`search_layer.cpp:on_delete`）
  —— 此前每次删除都做「LRU 拷原文 + NFKC + 分词」构造 `changed_terms`
  喂 `invalidate_terms`；但当 `cache_.size()==0`（写密集无查询负
  载）时后续 `invalidate_terms` 本就是 no-op，**重分词纯浪费**。
  改：空缓存走 `have_terms=true` 路径直接 `invalidate_terms(空)`
  —— 跳过 LRU 拷 + NFKC + 分词。LRU miss 仍降级 `cache_.invalidate()`
  （S9.2 安全粗粒度语义保留）。
* **P8.6/9：C `bitcask_get` 消除双拷贝**（`bitcask_c.cpp`）—— 此前
  路径：`get_owned`（span→vector 拷一次）+ `fill_get_result`（malloc
  + memcpy 再拷一次），每次 C get 多付整份 value 拷贝。新增
  `fill_get_result_view` 直接从零拷贝 `GetResultView` malloc + memcpy
  （语义同 `fill_get_result`，但接 view 而非 owned）。`bitcask_get`
  改 `casks->get()` 返回 `GetResultView`（零拷贝 view，page cache
 命中下经 mmap 直读）→ `fill_get_result_view`。
* **P8.11：FOR 解包改 64-bit 窗口**（`inverted.cpp:for_decode_block`）
  —— 此前每值 `bits` 次移位 + 掩码（v6 checkpoint 加载对每个 ord 付
  此代价，大索引启动主导项之一）。MSB-first 大端位流：8 字节大端
  一次 `byteswap` 载入（LE 主机已 `static_assert`），`off+bits` 内
  部窗口一次移位 + 掩码取值。`bits≤56`（实际 ord delta 位宽不超
  此限）走 64-bit 路径；越界（尾部不足 8 字节 / `bits>56`）回退
  逐 bit——位级等价。

未做的 P8 8 项（`search_fields` 死代码 + 逐词 top-k、`invalidate_terms`
反向索引、meta filter 锁内求值、merge `pending_` 分批 apply、hint
启动 CRC 边读边算、`rebuild_hnsw` 并行/结构化拷贝、`save_vec_payload`
流式、剩余 FOR 优化空间）按价值逐项评估落地——性价比从「meta filter
锁内求值」与「search_fields 死代码清理」最高。

CHANGELOG / TASK.md

* CHANGELOG [Unreleased] Performance 段新增 P7（seqlock）+ P8 6/14 各子
  项小注；既有 503/503 验证数字**不变**（本批零新单测，验证靠 bench
  + 既有 TSan 批）。
* TASK.md S13 段：C 组 P7 标完成（含 bench 数据 + 1→4 线程延迟持平
  实证）；P8 标 `[~]` 部分完成（6/14 列出，剩余 8 项登记跟进）。

验证

* Debug（clang）全量 503/503 不变（本批无新单测）；HNSW 15/15 既有
  单测含 recall 通过。
* TSan 并发相关全过：P7 seqlock 走 `atomic_ref` + release/acquire fence
  ——**无非原子冲突访问**，TSan 干净。P8.4 fuzzy 并行化 matchers 跨
  线程只读，与 wildcard 同一前提。
* P7 bench `BM_Hnsw_SearchConcurrent`：100k/ef64 × Threads 1/4/8 实测
  1→4 线程延迟持平（147→135µs），8 线程 CPU 时间不涨。
* Release 构建干净。

S13 现状：

| 组 | 进度 |
|---|---|
| A 并发 | ✅ 全完 |
| B 内存 | M1-M3 ✅，M4 按需 |
| C 性能 | ✅ **P1-P8 全完**（P8 6/14 部分 + 8 项后续） |
| D 功能 | ✅ 全完 |

+++ include/bitcask/hnsw.hpp
@@ -172,7 +172,13 @@ private:
         // adj_slabs，故 slab 增长对并发读者安全。
         std::vector<std::unique_ptr<std::uint32_t[]>> adj_slabs;
         std::size_t adj_slab_used = 0;
-        std::unique_ptr<std::atomic<std::uint8_t>[]> locks;  // per-node 自旋
+        // S13-P7：per-node seqlock 序号（偶=稳定，奇=写者更新中）。此前是
+        // 自旋锁——读者 copy_neighbors 对锁字节做 exchange（写操作），HNSW
+        // 流量高度偏向 hub 节点 → 并发查询时锁缓存行核间乒乓。写者是单
+        // 线程（reducer），无需互斥，只需发布协议：写者 ±1 包住 adj 更新，
+        // 读者双读序号一致才采信（torn 读被重试丢弃；数据字全走 atomic_ref
+        // relaxed，UB-free 且 TSan 干净）。u32 防 ABA 回绕。
+        std::unique_ptr<std::atomic<std::uint32_t>[]> locks;
 
         // V4.2:int8 量化副本(对称量化,scale = max |v[i]|)。codes 是紧
         // 排 int8,scale/sum_codes 是每向量一个标量;load/insert 时写入,
